+++ sentry-apollo-3/src/test/java/io/sentry/apollo3/SentryApollo3InterceptorTest.kt
@@ -1,14 +1,21 @@
 package io.sentry.apollo3
 
 import com.apollographql.apollo3.ApolloClient
+import com.apollographql.apollo3.api.http.HttpRequest
+import com.apollographql.apollo3.api.http.HttpResponse
 import com.apollographql.apollo3.exception.ApolloException
+import com.apollographql.apollo3.exception.ApolloHttpException
+import com.apollographql.apollo3.network.http.HttpInterceptor
+import com.apollographql.apollo3.network.http.HttpInterceptorChain
 import io.sentry.BaggageHeader
 import io.sentry.Breadcrumb
 import io.sentry.IHub
 import io.sentry.ITransaction
 import io.sentry.SentryOptions
+import io.sentry.SentryOptions.DEFAULT_PROPAGATION_TARGETS
 import io.sentry.SentryTraceHeader
 import io.sentry.SentryTracer
+import io.sentry.SpanDataConvention
 import io.sentry.SpanStatus
 import io.sentry.TraceContext
 import io.sentry.TracesSamplingDecision
@@ -40,7 +47,7 @@ class SentryApollo3InterceptorTest {
             whenever(options).thenReturn(
                 SentryOptions().apply {
                     dsn = "https://key@sentry.io/proj"
-                    isTraceSampling = true
+                    setTracePropagationTargets(listOf(DEFAULT_PROPAGATION_TARGETS))
                     sdkVersion = SdkVersion("test", "1.2.3")
                 }
             )
@@ -65,6 +72,7 @@ class SentryApollo3InterceptorTest {
   }
 }""",
             socketPolicy: SocketPolicy = SocketPolicy.KEEP_OPEN,
+            interceptor: HttpInterceptor? = null,
             addThirdPartyBaggageHeader: Boolean = false,
             beforeSpan: BeforeSpanCallback? = null
         ): ApolloClient {
@@ -83,6 +91,10 @@ class SentryApollo3InterceptorTest {
                 .serverUrl(server.url("/").toString())
                 .addHttpInterceptor(httpInterceptor)
 
+            interceptor?.let {
+                builder.addHttpInterceptor(interceptor)
+            }
+
             if (addThirdPartyBaggageHeader) {
                 builder.addHttpHeader("baggage", "thirdPartyBaggage=someValue")
                     .addHttpHeader("baggage", "secondThirdPartyBaggage=secondValue; property;propertyKey=propertyValue,anotherThirdPartyBaggage=anotherValue")
@@ -100,7 +112,7 @@ class SentryApollo3InterceptorTest {
 
         verify(fixture.hub).captureTransaction(
             check {
-                assertTransactionDetails(it)
+                assertTransactionDetails(it, httpStatusCode = 200)
                 assertEquals(SpanStatus.OK, it.spans.first().status)
             },
             anyOrNull<TraceContext>(),
@@ -115,8 +127,29 @@ class SentryApollo3InterceptorTest {
 
         verify(fixture.hub).captureTransaction(
             check {
-                assertTransactionDetails(it)
+                assertTransactionDetails(it, httpStatusCode = 403)
                 assertEquals(SpanStatus.PERMISSION_DENIED, it.spans.first().status)
+            },
+            anyOrNull<TraceContext>(),
+            anyOrNull(),
+            anyOrNull()
+        )
+    }
+
+    @Test
+    fun `get http status from ApolloHttpException in failed request`() {
+        val failingInterceptor = object : HttpInterceptor {
+            override suspend fun intercept(request: HttpRequest, chain: HttpInterceptorChain): HttpResponse {
+                throw ApolloHttpException(404, mock(), mock(), "")
+            }
+        }
+        executeQuery(fixture.getSut(interceptor = failingInterceptor))
+
+        verify(fixture.hub).captureTransaction(
+            check {
+                assertTransactionDetails(it, httpStatusCode = 404, contentLength = null)
+                assertEquals(404, it.spans.first().data?.get(SpanDataConvention.HTTP_STATUS_CODE_KEY))
+                assertEquals(SpanStatus.NOT_FOUND, it.spans.first().status)
             },
             anyOrNull<TraceContext>(),
             anyOrNull(),
@@ -130,7 +163,7 @@ class SentryApollo3InterceptorTest {
 
         verify(fixture.hub).captureTransaction(
             check {
-                assertTransactionDetails(it)
+                assertTransactionDetails(it, httpStatusCode = null, contentLength = null)
                 assertEquals(SpanStatus.INTERNAL_ERROR, it.spans.first().status)
             },
             anyOrNull<TraceContext>(),
@@ -239,8 +272,11 @@ class SentryApollo3InterceptorTest {
         verify(fixture.hub).addBreadcrumb(
             check<Breadcrumb> {
                 assertEquals("http", it.type)
-                assertEquals(280L, it.data["response_body_size"])
+                // response_body_size is added but mock webserver returns 0 always
+                assertEquals(0L, it.data["response_body_size"])
                 assertEquals(193L, it.data["request_body_size"])
+                assertEquals("LaunchDetails", it.data["operation_name"])
+                assertNotNull(it.data["operation_id"])
             },
             anyOrNull()
         )
@@ -255,13 +291,20 @@ class SentryApollo3InterceptorTest {
         assert(packageInfo.version == BuildConfig.VERSION_NAME)
     }
 
-    private fun assertTransactionDetails(it: SentryTransaction) {
+    private fun assertTransactionDetails(it: SentryTransaction, httpStatusCode: Int? = 200, contentLength: Long? = 0L) {
         assertEquals(1, it.spans.size)
         val httpClientSpan = it.spans.first()
         assertEquals("http.graphql", httpClientSpan.op)
         assertTrue { httpClientSpan.description?.startsWith("Post LaunchDetails") == true }
         assertNotNull(httpClientSpan.data) {
             assertNotNull(it["operationId"])
+            assertEquals("Post", it["http.method"])
+            httpStatusCode?.let { code ->
+                assertEquals(code, it[SpanDataConvention.HTTP_STATUS_CODE_KEY])
+            }
+            contentLength?.let { contentLength ->
+                assertEquals(contentLength, it[SpanDataConvention.HTTP_RESPONSE_CONTENT_LENGTH_KEY])
+            }
         }
     }